LPS-56451 extract xml formatting code to XMLUtil
