+++ modules/apps/journal/journal-test/test/integration/com/liferay/journal/service/test/BaseJournalServiceTestCase.java
@@ -21,7 +21,7 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.service.GroupLocalServiceUtil;
 import com.liferay.portal.service.ServiceContext;
-import com.liferay.portlet.dynamicdatamapping.util.DDMXMLUtil;
+import com.liferay.util.xml.XMLUtil;
 
 import java.io.InputStream;
 
@@ -57,7 +57,7 @@ public class BaseJournalServiceTestCase {
 	protected String getDefultXsd() throws Exception {
 		String xsd = readText("test-journal-structure-all-fields.xml");
 
-		return DDMXMLUtil.formatXML(xsd);
+		return XMLUtil.formatXML(xsd);
 	}
 
 	protected ServiceContext getServiceContext() {
