+++ modules/portal/portal-wab-generator/src/com/liferay/portal/wab/generator/internal/processor/WabProcessor.java
@@ -54,7 +54,7 @@ import com.liferay.portal.util.PropsValues;
 import com.liferay.portal.wab.generator.internal.introspection.ClassLoaderSource;
 import com.liferay.portal.wab.generator.internal.introspection.Source;
 import com.liferay.portal.wab.generator.internal.util.AntUtil;
-import com.liferay.portlet.dynamicdatamapping.util.DDMXMLUtil;
+import com.liferay.util.xml.XMLUtil;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -273,7 +273,7 @@ public class WabProcessor {
 		throws IOException {
 
 		try {
-			FileUtil.write(file, DDMXMLUtil.formatXML(document));
+			FileUtil.write(file, XMLUtil.formatXML(document));
 		}
 		catch (Exception e) {
 			throw new IOException(e);
