+++ portal-impl/src/com/liferay/portlet/dynamicdatamapping/service/impl/DDMTemplateLocalServiceImpl.java
@@ -50,6 +50,7 @@ import com.liferay.portlet.dynamicdatamapping.model.DDMTemplateConstants;
 import com.liferay.portlet.dynamicdatamapping.model.DDMTemplateVersion;
 import com.liferay.portlet.dynamicdatamapping.service.base.DDMTemplateLocalServiceBaseImpl;
 import com.liferay.portlet.dynamicdatamapping.util.DDMXMLUtil;
+import com.liferay.util.xml.XMLUtil;
 
 import java.io.File;
 import java.io.IOException;
@@ -1454,7 +1455,7 @@ public class DDMTemplateLocalServiceImpl
 				throw new TemplateScriptException(pe);
 			}
 
-			script = DDMXMLUtil.formatXML(script);
+			script = XMLUtil.formatXML(script);
 		}
 
 		return script;
