+++ util-java/src/com/liferay/util/xml/XMLUtil.java
@@ -1,0 +1,60 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.util.xml;
+
+import com.liferay.portal.kernel.exception.SystemException;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.xml.Document;
+
+import java.io.IOException;
+
+/**
+ * @author Leonardo Barros
+ */
+public class XMLUtil {
+
+	public static String formatXML(Document document) {
+		try {
+			return document.formattedString(_XML_INDENT);
+		}
+		catch (IOException ioe) {
+			throw new SystemException(ioe);
+		}
+	}
+
+	public static String formatXML(String xml) {
+
+		// This is only supposed to format your xml, however, it will also
+		// unwantingly change &#169; and other characters like it into their
+		// respective readable versions
+
+		try {
+			xml = StringUtil.replace(xml, "&#", "[$SPECIAL_CHARACTER$]");
+			xml = XMLFormatter.toString(xml, _XML_INDENT);
+			xml = StringUtil.replace(xml, "[$SPECIAL_CHARACTER$]", "&#");
+
+			return xml;
+		}
+		catch (IOException ioe) {
+			throw new SystemException(ioe);
+		}
+		catch (org.dom4j.DocumentException de) {
+			throw new SystemException(de);
+		}
+	}
+
+	private static final String _XML_INDENT = "  ";
+
+}
